code/requirements.txt
click==7.1.2
fastapi==0.63.0
h11==0.12.0
pydantic==1.8.1
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
yara-python==4.0.5
spdx-tools==0.6.1
